# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BreakEvenOut": {
        "properties": {
          "break_even_revenue": {
            "items": {
              "anyOf": [
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Break Even Revenue",
            "type": "array"
          },
          "margin_of_safety": {
            "items": {
              "anyOf": [
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Margin Of Safety",
            "type": "array"
          }
        },
        "required": [
          "break_even_revenue",
          "margin_of_safety"
        ],
        "title": "BreakEvenOut",
        "type": "object"
      },
...
      "ConsolidateResponse": {
        "description": "Сводный бюджет холдинга + разбивка вклада по проектам (B3).",
        "properties": {
          "actualized_cashflow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatementOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "balance": {
            "$ref": "#/components/schemas/StatementOut"
          },
          "break_even": {
            "$ref": "#/components/schemas/BreakEvenOut"
          },
          "cashflow": {
            "$ref": "#/components/schemas/StatementOut"
          },
          "cashflow_variance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatementOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "income": {
            "$ref": "#/components/schemas/StatementOut"
          },
          "metrics": {
            "$ref": "#/components/schemas/MetricsOut"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "per_project": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PerProjectOut"
            },
            "title": "Per Project",
            "type": "array"
          },
          "profit_use": {
            "$ref": "#/components/schemas/StatementOut"
          },
          "ratios": {
            "$ref": "#/components/schemas/RatiosOut"
          },
          "valuation": {
            "$ref": "#/components/schemas/ValuationOut"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "engine_version",
          "n",
          "income",
          "cashflow",
          "balance",
          "profit_use",
          "metrics",
          "ratios",
          "break_even",
          "valuation",
          "warnings"
        ],
        "title": "ConsolidateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HoldingConsolidationOut": {
        "description": "Сводка последней консолидации холдинга (B3).",
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "npv": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Npv",
            "type": "string"
          },
          "rate": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Rate",
            "type": "string"
          }
        },
        "required": [
          "npv",
          "rate",
          "at"
        ],
        "title": "HoldingConsolidationOut",
        "type": "object"
      },
      "HoldingCreate": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "HoldingCreate",
        "type": "object"
      },
      "HoldingMemberCreate": {
        "properties": {
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "role": {
            "default": "subsidiary",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "project_id"
        ],
        "title": "HoldingMemberCreate",
        "type": "object"
      },
      "HoldingMemberOut": {
        "properties": {
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "role"
        ],
        "title": "HoldingMemberOut",
        "type": "object"
      },
      "HoldingMemberPatch": {
        "properties": {
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "HoldingMemberPatch",
        "type": "object"
      },
      "HoldingOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_consolidation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HoldingConsolidationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "members": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HoldingMemberOut"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "HoldingOut",
        "type": "object"
      },
...
      "LineOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "values": {
            "items": {
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
              "type": "string"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "code",
          "label",
          "values"
        ],
        "title": "LineOut",
        "type": "object"
      },
...
      "MetricsOut": {
        "properties": {
          "dpb_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dpb Months"
          },
          "irr_annual": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irr Annual"
          },
          "npv": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Npv",
            "type": "string"
          },
          "pb_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb Months"
          },
          "peak_financing_need": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Financing Need"
          },
          "pi": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pi"
          },
          "pv_investments": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pv Investments"
          }
        },
        "required": [
          "npv"
        ],
        "title": "MetricsOut",
        "type": "object"
      },
...
      "PerProjectOut": {
        "description": "Вклад одного проекта в консолидацию (B3).",
        "properties": {
          "irr_annual": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irr Annual"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "net_profit_total": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Net Profit Total",
            "type": "string"
          },
          "npv": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Npv",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revenue_total": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Revenue Total",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "role",
          "npv",
          "revenue_total",
          "net_profit_total"
        ],
        "title": "PerProjectOut",
        "type": "object"
      },
...
      "RatiosOut": {
        "properties": {
          "activity": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Activity",
            "type": "object"
          },
          "gearing": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Gearing",
            "type": "object"
          },
          "investment": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Investment",
            "type": "object"
          },
          "liquidity": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Liquidity",
            "type": "object"
          },
          "profitability": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Profitability",
            "type": "object"
          }
        },
        "required": [
          "liquidity",
          "activity",
          "gearing",
          "profitability",
          "investment"
        ],
        "title": "RatiosOut",
        "type": "object"
      },
...
      "StatementOut": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineOut"
            },
            "title": "Lines",
            "type": "array"
          }
        },
        "required": [
          "lines"
        ],
        "title": "StatementOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Расчёт финансовой модели предприятия (отчёты, показатели, коэффициенты).",
    "title": "Финансовая модель — API",
    "version": "0.9.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/holdings": {
      "get": {
        "operationId": "list_holdings_api_v1_holdings_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HoldingOut"
                  },
                  "title": "Response List Holdings Api V1 Holdings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Holdings",
        "tags": [
          "holdings"
        ]
      },
      "post": {
        "operationId": "create_holding_api_v1_holdings_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Holding",
        "tags": [
          "holdings"
        ]
      }
    },
    "/api/v1/holdings/{holding_id}": {
      "delete": {
        "operationId": "delete_holding_api_v1_holdings__holding_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "holding_id",
            "required": true,
            "schema": {
              "title": "Holding Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Holding",
        "tags": [
          "holdings"
        ]
      },
      "get": {
        "operationId": "get_holding_api_v1_holdings__holding_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "holding_id",
            "required": true,
            "schema": {
              "title": "Holding Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Holding",
        "tags": [
          "holdings"
        ]
      }
    },
    "/api/v1/holdings/{holding_id}/consolidate": {
      "post": {
        "description": "Сводный бюджет холдинга (консолидация) + разбивка вклада по проектам (B3).",
        "operationId": "consolidate_holding_api_v1_holdings__holding_id__consolidate_post",
        "parameters": [
          {
            "in": "path",
            "name": "holding_id",
            "required": true,
            "schema": {
              "title": "Holding Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "group_discount_rate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ],
              "default": "0.15",
              "title": "Group Discount Rate"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsolidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consolidate Holding",
        "tags": [
          "holdings"
        ]
      }
    },
    "/api/v1/holdings/{holding_id}/members": {
      "post": {
        "operationId": "add_member_api_v1_holdings__holding_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "holding_id",
            "required": true,
            "schema": {
              "title": "Holding Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldingMemberCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Member",
        "tags": [
          "holdings"
        ]
      }
    },
    "/api/v1/holdings/{holding_id}/members/{project_id}": {
      "delete": {
        "description": "Исключить проект из холдинга (сам проект не удаляется).",
        "operationId": "remove_member_api_v1_holdings__holding_id__members__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "holding_id",
            "required": true,
            "schema": {
              "title": "Holding Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Member",
        "tags": [
          "holdings"
        ]
      },
      "patch": {
        "description": "Изменить роль участника холдинга (parent | subsidiary).",
        "operationId": "patch_member_role_api_v1_holdings__holding_id__members__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "holding_id",
            "required": true,
            "schema": {
              "title": "Holding Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldingMemberPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Member Role",
        "tags": [
          "holdings"
        ]
      }
    },
...
  }
}